LPS-179214 Extract ScreenNavigationEntry methods from AccountEntryAccountGroupsScreenNavigationCategory to AccountEntryAccountGroupsScreenNavigationEntry, implement getCategoryKey method in it and return getCategoryKey in getEntryKey method
